+++ Source/LGUI/Private/Core/ActorComponent/UIBackgroundPixelate.cpp
@@ -108,7 +108,7 @@ public:
 		const FMatrix44f& ViewProjectionMatrix,
 		bool IsWorldSpace,
 		float BlendDepthForWorld,
-		float DepthFadeForWorld,
+		int DepthFadeForWorld,
 		const FIntRect& ViewRect,
 		const FVector4f& DepthTextureScaleOffset,
 		const FVector4f& ViewTextureScaleOffset
