+++ Source/LGUI/Public/Core/UIPostProcessRenderProxy.h
@@ -43,7 +43,7 @@ public:
 		const FMatrix44f& ViewProjectionMatrix,
 		bool IsWorldSpace,
 		float BlendDepthForWorld,
-		float DepthFadeForWorld,
+		int DepthFadeForWorld,
 		const FIntRect& ViewRect,
 		const FVector4f& DepthTextureScaleOffset,
 		const FVector4f& ViewTextureScaleOffset
@@ -76,7 +76,7 @@ public:
 		, const FMatrix44f & ModelViewProjectionMatrix
 		, bool IsWorldSpace
 		, float BlendDepthForWorld
-		, float DepthFadeForWorld
+		, int DepthFadeForWorld
 		, const FVector4f& DepthTextureScaleOffset
 		, const FIntRect& ViewRect
 		, FRHISamplerState* ResultTextureSamplerState = TStaticSamplerState<SF_Bilinear, AM_Clamp, AM_Clamp, AM_Clamp>::GetRHI()
